requirements.txt
python-dotenv~=1.0.1
requests~=2.32.3
pandas~=2.2.3
deep-translator~=1.11.4
translate~=3.6.1